requirements.txt
Flask~=2.1.1
python-dotenv~=0.20.0
gunicorn~=20.1.0